Add conditional halo pass for frazil inside make_frazil (#932)
* Add conditional halo pass for frazil inside make_frazil

Frazil may need an extra halo pass, but only when all of the following are true:
(1) frazil is calculated and applied in the halo (e.g., vertex_shear is true)
(2) the thermodynamic loop is called more than once before passing frazil to the sea-ice (dt_therm<dt_cpld)
(3) reclaim_frazil is true
(4) a previous make_frazil was called after the diabatic loop and before T&S were updated in halos.

This commit adds the halo pass inside make_frazil since it is where we can know all of these conditions at once.  We needed to add a flag to know if frazil was reset (which happens if dt_therm>=dt_cpld), in which case the halo pass would be unnecessary.  We needed an optional argument for make_frazil so we could flag to only do this update on the first call to make_frazil (which is needed because the previous call to make_frazil was after the diabatic loop, condition 4)

* Shorten line in MOM_diabatic_driver

* Move halo update on frazil out of make_frazil and into post_diabatic_halo_updates

---------

Co-authored-by: brandon.reichl <brandon.reichl@noaa.gov>