Horvisc: Invert do/if blocks of better_bound
do concurrent / if / elseif / enddo gave different answers for some
reason.  Compiler bug?  Each iteration should be thread-safe.  Maybe
it's a compiler bug.

Anyway, moving the if-blocks outside of do-concurrent restored
reproducibility with CPU.  Results now agree for Kh_q.

The Kh_q was loop also moved back to CPU.  Not necessarily required, but
it was done during troubleshooting.  Kh_h is on CPU so why not?  (This
will all be revisted someday.)