Barotropic: Checksum cleanup
This patch moves any checksummed fields from GPU to CPU before
calculation.  This helps to sort out any diffs in the stderr logs.

This is a bit tedious, and it might make more sense to move this into
the actual checksum function.  But currently not every field will be on
the GPU, and a GPU->CPU transfer could actually ruin the checksum.

For now, we just do this in the usual tedious manner.