Barotropic: CS mem cleanup; remove btcalc copies
This patch replaces the safer barotropic CS alloc/delete inside of
btcalc and replaces it with a persistent CS created by barotropic_init.

A difficulty here is that btcalc is called by both the dynamic core step
function (e.g. stype_dyn_split_RK2) as well as barotropic_init, meaning
that it needs to be partially set up before btcalc can be called to
finialize the barotropic setup.

This is mostly a matter of timing each event so that the GPU state is
not out of sync with any of the CPU operations.

I'm not going to pretend that I understand everything here.  Multiple
attempts failed, and this is the one that worked.  We still have some
lessons to learn about CS initialization.