DO_LOCALITY() bugfix; use GPU FMS in .testing
The default FMS is not compatible with the new GPU-based MPI methods, so
we just change it in .testing/Makefile for now.  Down the road, we
should probably move this into the .github config.

DO_LOCALITY() now returns a ; if do concurrent locality modifiers are
unsupported.  This prevents errors associated with line continuations.